k8s_projects/tiktok-transcriber/backend/requirements.txt
fastapi
uvicorn
yt-dlp
ffmpeg-python
python-multipart
openai-whisper
